% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA2.R
\name{plotPCA2}
\alias{plotPCA2}
\title{Representant PCAs}
\usage{
plotPCA2(X, labels = NULL, colors = NULL, dataDesc = "", scale = FALSE,
  transpose = FALSE)
}
\arguments{
\item{X}{El nom de la matriu de la que es calculara el PCA (de t(X))}

\item{labels}{Vector d'etiquetes per a la visualització dels punts. Per defecte es NULL}

\item{colors}{Vector de colors per als punts. De la mateixa mida que el de les etiquetes.}

\item{dataDesc}{Breu descripcio de les dades per al titol}

\item{scale}{Variable logica per decidir si s'escalen les dades}

\item{transpose}{Variable logica per decidir si es vol fer el PCA de la matriu original o de la seva trasposta}
}
\description{
Representant PCAs
}
\keyword{Multivariate,}
\keyword{Plots}

