% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA2_function.R
\name{plotPCA2}
\alias{plotPCA2}
\title{Plot PCA}
\usage{
plotPCA2(datos, labels, factor, title, scale)
}
\arguments{
\item{datos}{A data set where only the first column is character (usually sample names)}

\item{labels}{A vector with the sample names (usually in targets$shortname column)}

\item{factor}{A vector with the different levels to study in the PCA (usually in targets$Grupo column)}

\item{title}{The name of the factor which you are colouring the sample}

\item{scale}{The data needs to be scaled? Values= TRUE/FALSE}
}
\description{
This function plot the two first principal components of a dataset
}
\examples{
plotPCA2(datos=ExampleData$data,labels = ExampleData$targets$shortname,
factor=ExampleData$targets$grupo,title = "Grupo",scale=FALSE)
}

